// src/app/model/record.dto.ts
export interface RecordDto {
     /**
   * 紀錄的唯一識別碼
   * - 後端可能給定，也可能尚未給定（例如剛建立資料時）
   * - 可選欄位，表示這筆紀錄可能沒有 `id`
   * - 系統／後端負責產生並回傳這個值
   */
  id?: string | null;

  /**
   * 使用者 ID
   * - 表示這筆收支紀錄屬於哪位使用者
   * - 必填欄位
   */
  userId: string;

  /**
   * 日期
   * - 可以是 ISO 8601 格式字串，也可以是 YYYY-MM-DD 簡寫
   * - 表示這筆收支發生的日期（也可能含時間）
   */
  date: string ;

  /**
   * 類型
   * - 'income' 表示收入
   * - 'expense' 表示支出
   * - 用來區分這筆紀錄是收入還是支出類型
   */
  type: 'income' | 'expense';

  /**
   * 金額
   * - 必填欄位
   * - 使用正數或小數來表示這筆紀錄的金額
   */
  amount: number;

  /**
   * 類別 ID
   * - 對應某個收支分類（例如「餐飲」「交通」「薪水」等）
   * - 用來做分類與篩選用途
   */
  categoryId: string;

  /**
   * 描述／備註
   * - 可選欄位
   * - 使用者可補充文字來說明這筆記錄用途或細節
   */
  description?: string |null ;

  /**
   * 支付方式
   * - 可選欄位
   * - 表示這筆支出／收入是以什麼方式付／收款，例如：現金、信用卡、行動支付、銀行轉帳等
   */
  paymentMethod: string ;
}